Added logging event tests

+++ worker/src/main/java/gov/cms/ab2d/worker/processor/JobProcessorImpl.java
@@ -145,7 +145,7 @@ public class JobProcessorImpl implements JobProcessor {
             eventLogger.log(new ContractBeneSearchEvent(job.getUser() == null ? null : job.getUser().getUsername(),
                     job.getJobUuid(),
                     contract.getContractNumber(),
-                    (int) progressTracker.getContractCount(contract.getContractNumber()),
+                    progressTracker.getContractCount(contract.getContractNumber()),
                     progressTracker.getProcessedCount(),
                     progressTracker.getOptOutCount(),
                     progressTracker.getFailureCount()));
